\name{mtsc}
\alias{mtsc}
\alias{summary.mtsc}
\title{place holder function}
\description{Early prototypes.}
\usage{
mtsc (t, mts, fit=TRUE, standardise=TRUE)
\method{summary}{mtsc}(m, \dots)
}
\arguments{
\item{t}{A vector of time values.}
\item{mts}{A Matrix of response values.}
\item{fit}{Fit the model.}
\item{standardise}{Standardise the matrix.}
\item{m}{A mtsc model.}
\item{\dots}{Ignored.}
}


